Expose the attachments for Assessor.
Exposed attachment resource for the Assessor. Extracted the module with shared methods to reuse it in different controller namespaces. Probably something to discuss as not very popular in Rails world but looks as easy way to achieving non dry controllers for Assessor/Admin spaces.
