+++ app/controllers/assessor/form_answer_attachments_controller.rb
@@ -1,0 +1,3 @@
+class Assessor::FormAnswerAttachmentsController < Assessor::BaseController
+  include ::FormAnswerAttachmentsContext
+end
