+++ app/policies/form_answer_attachment_policy.rb
@@ -1,4 +1,6 @@
 class FormAnswerAttachmentPolicy < ApplicationPolicy
+  # TODO: needs clarification
+
   def create?
     true
   end
